backend/app/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-dotenv>=0.19.1
google-generativeai>=0.3.0
httpx>=0.23.0